+++ frontend/app.vue
@@ -1,16 +1,5 @@
 <script setup lang="ts">
-import { serializeError } from 'serialize-error'
-
-const { loading: { clearLoading }, error: { errorModalVisible, error } } = useModal()
-
-const handleError = (err: Error) => {
-  if (process.client) {
-    console.log(err)
-  }
-  error.value = serializeError(err)
-  errorModalVisible.value = true
-  clearLoading()
-}
+const { error: { handleError } } = useModal()
 
 onErrorCaptured((err: unknown, _instance: ComponentPublicInstance | null, _info: string) => {
   let error: Error | undefined
